添加: 基于HandlerChain的自定义消息处理流程 调整: 重构rpc处理流程 移除: Session.Post 移除: DispatcherHandler 移除: Read/WriteSocketHandler

+++ proto/binary/coredef/binmsg.go
@@ -38,7 +38,7 @@ func (m *SessionClosed) String() string { return goobjfmt.CompactTextString(m) }
 type RemoteCallACK struct {
 	MsgID  uint32
 	Data   []byte
-	CallID int32
+	CallID int64
 }
 
 func (m *RemoteCallACK) String() string { return goobjfmt.CompactTextString(m) }
